+++ crypto-scraper/app/Console/Commands/Bitcointalk/Loaders/UrlValidations.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Console\Commands\Bitcointalk\DBLoaders;
+namespace App\Console\Commands\Bitcointalk\Loaders;
 
 
 trait UrlValidations {
