+++ crypto-scraper/tests/Unit/Bitcointalk/GetBoardPagesTest.php
@@ -3,8 +3,8 @@ declare(strict_types=1);
 
 namespace Tests\Unit\Bitcointalk;
 
-use App\Console\Commands\Bitcointalk\DBLoaders\Boards as LoadBoards;
-use App\Console\Commands\Bitcointalk\DBLoaders\UrlValidations;
+use App\Console\Commands\Bitcointalk\Loaders\Boards as LoadBoards;
+use App\Console\Commands\Bitcointalk\Loaders\UrlValidations;
 use Tests\TestCase;
 
 class GetBoardPagesTest extends TestCase {
