+++ crypto-scraper/tests/Unit/Bitcointalk/MainBoardValidTest.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace Tests\Unit\Bitcointalk;
 
-use App\Console\Commands\Bitcointalk\DBLoaders\Boards as LoadBoards;
+use App\Console\Commands\Bitcointalk\Loaders\Boards as LoadBoards;
 use Tests\TestCase;
 
 class MainBoardValidTest extends TestCase
